Extract a base runBlocking liveness test
